feat: add channel module. refactor app struct
