+++ src/app/app-routing.module.ts
@@ -1,0 +1,11 @@
+import {NgModule} from "@angular/core";
+import {RouterModule, Routes} from "@angular/router";
+
+const routes: Routes = [
+]
+
+@NgModule({
+  imports: [RouterModule.forRoot(routes)],
+  exports: [RouterModule],
+})
+export class AppRoutingModule {}
