+++ src/app/app.component.html
@@ -1,1 +1,1 @@
-<app-chanells></app-chanells>
+<router-outlet></router-outlet>
