+++ src/app/app.module.ts
@@ -1,25 +1,22 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import {RouterModule, Routes} from "@angular/router";
 import {HttpClientModule} from "@angular/common/http";
-import {MatCardModule} from '@angular/material/card';
-
 import { AppComponent } from './app.component';
-import {ChannelsComponent} from "./modules/chanels_module/channels/channels.component";
-import {ChannelCardComponent} from "./modules/chanels_module/chanel-card/channel-card.component";
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+
+import {ChannelsModule} from "./modules/channels.module";
+import {AppRoutingModule} from "./app-routing.module";
 
 @NgModule({
   declarations: [
     AppComponent,
-    ChannelsComponent,
-    ChannelCardComponent,
   ],
   imports: [
     BrowserModule,
+    AppRoutingModule,
+    ChannelsModule,
     HttpClientModule,
     BrowserAnimationsModule,
-    MatCardModule,
   ],
   providers: [],
   bootstrap: [AppComponent]
