+++ src/app/modules/chanels_module/components/chanel-card/channel-card.component.html
